Extract "imports" collection logic from SwiftGenerator
Extracted logic for collecting "imports" from SwiftGenerator into a stand-alone class SwiftImportsCollector.

See: #810
