+++ gluecodium/src/main/java/com/here/gluecodium/generator/swift/SwiftImportsResolver.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2016-2021 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.gluecodium.generator.swift
+
+import com.here.gluecodium.generator.common.ImportsResolver
+import com.here.gluecodium.model.lime.LimeElement
+import com.here.gluecodium.model.lime.LimeExternalDescriptor
+import com.here.gluecodium.model.lime.LimeNamedElement
+
+internal class SwiftImportsResolver : ImportsResolver<String> {
+    override fun resolveElementImports(limeElement: LimeElement) =
+        listOfNotNull(
+            (limeElement as? LimeNamedElement)?.external?.swift?.get(LimeExternalDescriptor.FRAMEWORK_NAME)
+        ).filter { it.isNotBlank() }
+}
